Refactor RecordList component structure
- Removed RecordListContext and ViewToggle components.
- Introduced RecordTableView component for displaying records in a table format.
- Consolidated configuration settings into a single config file.
- Added hooks for managing view preferences and column visibility.
- Updated index files to reflect new component structure and exports.
