--- src/components/RecordList/components/index.js
@@ -1,2 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export { default as ViewToggle } from './ViewToggle';
