--- src/components/RecordList/config/index.js
@@ -1,4 +1,0 @@
-export { reviewerConfig } from './reviewerConfig';
-export { publishedConfig } from './publishedConfig';
-export { submissionsConfig } from './submissionsConfig';
-export * from './columnDefinitions';
