+++ src/components/RecordList/context.js
@@ -1,0 +1,18 @@
+import { createContext, useContext } from 'react';
+
+export const RecordListContext = createContext({
+  config: null,
+  actionHandlers: {},
+  language: 'en',
+  region: '',
+});
+
+export const RecordListProvider = RecordListContext.Provider;
+
+export const useRecordListContext = () => {
+  const context = useContext(RecordListContext);
+  if (!context) {
+    throw new Error('useRecordListContext must be used within a RecordListProvider');
+  }
+  return context;
+};
