--- src/components/RecordList/hooks/index.js
@@ -1,3 +1,0 @@
-export { useViewPreference } from './useViewPreference';
-export { useColumnVisibility } from './useColumnVisibility';
-export { useResponsiveDefault } from './useResponsiveDefault';
