--- src/components/RecordList/hooks/useColumnVisibility.js
@@ -1,34 +1,0 @@
-import { useState, useCallback } from 'react';
-
-export function useColumnVisibility(storageKey, defaultVisibility) {
-  const [columnVisibilityModel, setColumnVisibilityModel] = useState(() => {
-    try {
-      const saved = localStorage.getItem(storageKey);
-      if (saved) {
-        return JSON.parse(saved);
-      }
-    } catch (e) {
-      // Ignore errors
-    }
-    return defaultVisibility;
-  });
-
-  const handleColumnVisibilityChange = useCallback(
-    (newModel) => {
-      setColumnVisibilityModel(newModel);
-      try {
-        localStorage.setItem(storageKey, JSON.stringify(newModel));
-      } catch (e) {
-        // Ignore errors
-      }
-    },
-    [storageKey]
-  );
-
-  return {
-    columnVisibilityModel,
-    handleColumnVisibilityChange,
-  };
-}
-
-export default useColumnVisibility;
