+++ src/components/RecordList/index.js
@@ -1,6 +1,4 @@
 export { default } from './RecordList';
-export { default as RecordList } from './RecordList';
-export { RecordListProvider, useRecordListContext } from './RecordListContext';
-export { reviewerConfig, publishedConfig, submissionsConfig } from './config';
+export { default as RecordList, RecordListProvider, useRecordListContext } from './RecordList';
+export { reviewerConfig, publishedConfig, submissionsConfig, createColumns, recordToRow } from './config';
 export { useViewPreference, useColumnVisibility, useResponsiveDefault } from './hooks';
-export { ViewToggle } from './components';
