--- src/components/RecordList/views/index.js
@@ -1,2 +1,0 @@
-export { default as RecordTableView } from './RecordTableView';
-export { default as RecordCardView } from './RecordCardView';
